Refactor: move bipartite algorithms to a separate module
