+++ test/Algebra/Graph/Test/Arbitrary.hs
@@ -24,20 +24,21 @@ import Algebra.Graph
 import Algebra.Graph.Export
 import Algebra.Graph.Label
 
-import qualified Algebra.Graph.Acyclic.AdjacencyMap   as AAM
-import qualified Algebra.Graph.AdjacencyIntMap        as AIM
-import qualified Algebra.Graph.AdjacencyMap           as AM
-import qualified Algebra.Graph.Bipartite.AdjacencyMap as BAM
-import qualified Algebra.Graph.NonEmpty.AdjacencyMap  as NAM
-import qualified Algebra.Graph.Class                  as C
-import qualified Algebra.Graph.Labelled               as LG
-import qualified Algebra.Graph.Labelled.AdjacencyMap  as LAM
-import qualified Algebra.Graph.NonEmpty               as NonEmpty
-import qualified Algebra.Graph.Relation               as Relation
-import qualified Algebra.Graph.Relation.Preorder      as Preorder
-import qualified Algebra.Graph.Relation.Reflexive     as Reflexive
-import qualified Algebra.Graph.Relation.Symmetric     as Symmetric
-import qualified Algebra.Graph.Relation.Transitive    as Transitive
+import qualified Algebra.Graph.Acyclic.AdjacencyMap             as AAM
+import qualified Algebra.Graph.AdjacencyIntMap                  as AIM
+import qualified Algebra.Graph.AdjacencyMap                     as AM
+import qualified Algebra.Graph.Bipartite.AdjacencyMap           as BAM
+import qualified Algebra.Graph.Bipartite.AdjacencyMap.Algorithm as BAMA
+import qualified Algebra.Graph.NonEmpty.AdjacencyMap            as NAM
+import qualified Algebra.Graph.Class                            as C
+import qualified Algebra.Graph.Labelled                         as LG
+import qualified Algebra.Graph.Labelled.AdjacencyMap            as LAM
+import qualified Algebra.Graph.NonEmpty                         as NonEmpty
+import qualified Algebra.Graph.Relation                         as Relation
+import qualified Algebra.Graph.Relation.Preorder                as Preorder
+import qualified Algebra.Graph.Relation.Reflexive               as Reflexive
+import qualified Algebra.Graph.Relation.Symmetric               as Symmetric
+import qualified Algebra.Graph.Relation.Transitive              as Transitive
 
 -- | Generate an arbitrary 'C.Graph' value of a specified size.
 arbitraryGraph :: (C.Graph g, Arbitrary (C.Vertex g)) => Gen g
@@ -248,5 +249,5 @@ instance (Arbitrary a, Arbitrary b) => Arbitrary (BAM.List a b) where
                   s <- arbitrary
                   (BAM.Cons f . BAM.Cons s) <$> go (n - 2)
 
-instance (Arbitrary a, Arbitrary b, Ord a, Ord b) => Arbitrary (BAM.Matching a b) where
-    arbitrary = BAM.matching <$> arbitrary
+instance (Arbitrary a, Arbitrary b, Ord a, Ord b) => Arbitrary (BAMA.Matching a b) where
+    arbitrary = BAMA.matching <$> arbitrary
